third_party/blink/web_tests/flag-specific/disable-site-isolation-trials/wpt_internal/svg/use-count-svg-filter-remote-iframe.sub.https-expected.txt

This is a testharness.js-based test.
[FAIL] SVG filter applied to remote frame
  assert_equals: Outer filter. expected "Local: false, Remote: true, Plugin: false" but got "Local: true, Remote: false, Plugin: false"
[FAIL] CSS & SVG filter applied to remote frame
  assert_equals: Outer filter SVG. expected "Local: false, Remote: true, Plugin: false" but got "Local: true, Remote: false, Plugin: false"
Harness: the test ran to completion.
